🌱 Split ServiceDrillDown.tsx into orchestrator, parts, and hook
Extracts ServiceDrillDown.tsx (480 lines, 11 hook calls) following the
established X.tsx + X.parts.tsx + useX.ts drilldown-split pattern.

- useServiceDrillDown.ts — all state (service type, IPs, ports,
  endpoints, describe/yaml output), the kubectl fetch effect, and
  the copy-to-clipboard handler
- ServiceDrillDown.parts.tsx — pure presentational sub-components
  (header, tab bar, overview/endpoints tabs, InfoField, shared
  describe/yaml output pane)
- ServiceDrillDown.tsx — thin orchestrator (101 lines, 3 hook calls)

No behaviour change — pure extraction. Default export and drill-down
registration unchanged.
